backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
playwright==1.41.2
pytz==2024.1
pydantic==2.5.2 